+++ src/script_interpreter_func_fe3d_client_get.cpp
@@ -79,9 +79,9 @@ const bool ScriptInterpreter::_executeFe3dClientGetter(const string& functionNam
 	{
 		if(_validateArgumentCount(args, 0) && _validateArgumentTypes(args, {}))
 		{
-			const auto result = _fe3d->client_getPingLatency();
+			const auto result = static_cast<int>(_fe3d->client_getPingLatency());
 
-			returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, static_cast<int>(result)));
+			returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 		}
 	}
 	else if(functionName == "fe3d:client_get_server_ip")
